Change ser

+++ HHAzureImageStorage/HHAzureImageStorage.Core/HHAzureImageStorage.Core/Models/Responses/ImageResizeResponse.cs
@@ -10,6 +10,6 @@ namespace HHAzureImageStorage.Core.Models.Responses
 
         public int HeightPixels { get; set; }
 
-        public int SizeInBytes { get; set; }
+        public long SizeInBytes { get; set; }
     }
 }
