+++ HHAzureImageStorage/HHAzureImageStorage.Domain/Entities/Image.cs
@@ -33,6 +33,6 @@ namespace HHAzureImageStorage.Domain.Entities
 
         public int HeightPixels { get; set; }
 
-        public int SizeInBytes { get; set; }
+        public long SizeInBytes { get; set; }
     }
 }
